+++ app/src/main/java/com/example/unserhoersaal/model/DatabaseCourses.java
@@ -12,7 +12,6 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 
 
 public class DatabaseCourses {
@@ -52,5 +51,37 @@ public class DatabaseCourses {
 
     public MutableLiveData<ArrayList<UserCourse>> getUserCourses() {
         return userCourses;
+    }
+
+    public static class Message {
+        public String messageText;
+        public Long time;
+
+
+        public Message() {
+            // Default constructor required for calls to DataSnapshot.getValue(User.class)
+        }
+
+        public Message(String messageText, Long time) {
+            this.messageText = messageText;
+            this.time = time;
+
+        }
+
+        public Long getTime() {
+            return time;
+        }
+
+        public String getMessageText() {
+            return messageText;
+        }
+
+        public void setMessageText(String messageText) {
+            this.messageText = messageText;
+        }
+
+        public void setTime(Long time) {
+            this.time = time;
+        }
     }
 }
