+++ app/src/main/java/com/example/unserhoersaal/views/ChatAdapter.java
@@ -8,14 +8,15 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 import androidx.recyclerview.widget.RecyclerView;
 import com.example.unserhoersaal.R;
-import com.example.unserhoersaal.model.Message;
+import com.example.unserhoersaal.model.DatabaseCourses;
+
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
 /** Chatadapter. */
 public class ChatAdapter extends RecyclerView.Adapter<ChatAdapter.ViewHolder> {
 
-  private Message[] localDataSet;
+  private DatabaseCourses.Message[] localDataSet;
   public static final SimpleDateFormat RECENT_FORMAT = new SimpleDateFormat("HH:mm");
   public static final SimpleDateFormat OLD_FORMAT = new SimpleDateFormat("dd. MMMM");
 
@@ -42,7 +43,7 @@ public class ChatAdapter extends RecyclerView.Adapter<ChatAdapter.ViewHolder> {
   }
 
 
-  public ChatAdapter(Message[] dataSet) {
+  public ChatAdapter(DatabaseCourses.Message[] dataSet) {
     localDataSet = dataSet;
   }
